requirements.txt
Django==6.0.3
django-extensions==4.1
djangorestframework==3.16.1
python-decouple==3.8
django-jazzmin==3.0.3
drf-material==0.3.4
icecream==2.1.10
uvicorn[standard]==0.41.0
gunicorn==25.1.0
whitenoise==6.12.0
django-debug-toolbar==6.2.0
django-cors-headers==4.9.0
django-import-export[all]==4.4.0
django-schema-viewer==0.5.3
drf-spectacular[sidecar]==0.29.0
django-filter==25.2
djangorestframework-jsonapi==8.1.0
Pillow==12.1.1
django-unfold==0.83.1
django-celery-beat==2.9.0
django-celery-results==2.6.0
dj-database-url==3.1.2
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1
django-ratelimit==4.1.0
cryptography==46.0.5
drf-standardized-errors[openapi]==0.15.0
django-parler==2.3
django-parler-rest==2.2
django-modeltranslation==0.20.2
django-rosetta==0.10.3
adrf==0.1.12
celery[redis,rabbitmq]==5.6.2
flower==2.0.1
django-daisy==2.0.8
django-redis==6.0.0
Faker==40.8.0
drf-spectacular-sidecar==2026.3.1